// app/_utils/types.ts
export enum VitalType {
  BLOOD_PRESSURE = 'Blood Pressure',
  HEART_RATE = 'Heart Rate',
  TEMPERATURE = 'Temperature',
  BLOOD_OXYGEN = 'Blood Oxygen',
  WEIGHT = 'Weight',
  HEIGHT = 'Height',
  BLOOD_SUGAR = 'Blood Sugar',
  PULSE_RATE = "Pulse Rate",
  BMI = 'BMI'
}